+++ client/services/TaskService.tsx
@@ -1,8 +1,161 @@
 import { getRoute } from '@/modules/map/MapService';
 import { calculateEuclideanDistance } from '@/modules/map/MapUtils';
-import { Route, Task, Location } from '@/types';
+import { Route, Task, Location, Position } from '@/types';
+import { generateMissingDurations, generateMissingLocations } from './gptService';
 
 export class TaskService {
+  static async generateTaskRoute(
+    startLocation: Location,
+    tasks: Task[]
+  ): Promise<{
+    route: Route;
+    tasks: Task[];
+  }> {
+    const coreTaskCandidates = tasks.filter((task) => task.startTime !== null);
+    let fillerTasks = tasks.filter((task) => task.startTime === null);
+
+    const generateDurationsPromise = generateMissingDurations(tasks);
+    const generateLocationsPromise = generateMissingLocations(tasks, [
+      startLocation.coordinates[0],
+      startLocation.coordinates[1],
+    ]);
+
+    // Generating missing durations and locations for all tasks
+    await Promise.all([generateDurationsPromise, generateLocationsPromise]);
+
+    // Core tasks have a start time and a location - tasks with a set time
+    // and with no location at this point are implicitly discarded
+    let coreTasks = coreTaskCandidates
+      .filter((task) => task.location !== null)
+      .sort((a, b) => {
+        const aHour = a.startTime?.getHours() || 0;
+        const bHour = b.startTime?.getHours() || 0;
+        return aHour - bHour;
+      });
+
+    let partialRoutes = await TaskService.getOptimalRouteForPaths(startLocation, coreTasks);
+
+    // START FILLING IN FILLER TASKS
+    const updatedPartialRoutes: Route[] = [];
+    const updatedCoreTasks: Task[] = [];
+    const usedFillerTaskIds = new Set<string>();
+    fillerTasks = fillerTasks.filter((task) => task.location !== null); // drop filler tasks with no location
+    for (let i = 0; i < partialRoutes.length; i++) {
+      const startTask =
+        i === 0
+          ? { duration: 0, startTime: new Date(), location: startLocation }
+          : coreTasks[i - 1];
+      const endTask = coreTasks[i];
+      const route = partialRoutes[i];
+
+      const maximumFreeTime =
+        (endTask.startTime?.getTime() ?? 0) -
+        ((startTask.startTime?.getTime() ?? 0) + (startTask.duration ?? 0) * 60 * 1000) -
+        route.duration * 1000;
+
+      // Obtain all combinations of filler tasks that fit within the maximum free time
+      const unusedFillerTasks = fillerTasks.filter((task) => !usedFillerTaskIds.has(task.id));
+      const feasibleFillerTaskCombinations: Task[][] = [];
+      for (const taskCombination of TaskService.allCombinations<Task>(unusedFillerTasks)) {
+        if (taskCombination.length === 0) continue;
+        const totalDuration = taskCombination.reduce((acc, task) => acc + (task.duration ?? 0), 0);
+        if (totalDuration * 60 * 1000 <= maximumFreeTime) {
+          feasibleFillerTaskCombinations.push(taskCombination);
+        }
+      }
+
+      // Order each combination to minimize travel distance
+      const orderedFillerTaskCombinations: { distance: number; tasks: Task[] }[] =
+        feasibleFillerTaskCombinations.map((combination) =>
+          TaskService.travelDistanceAwareReOrder(
+            combination,
+            startTask.location!,
+            endTask.location!
+          )
+        );
+
+      // Order the combinations themselves by task count (decreasing) and travel distance (increasing)
+      const orderedFillerTaskCombinationsSorted = orderedFillerTaskCombinations.sort((a, b) => {
+        const taskCountDiff = b.tasks.length - a.tasks.length; // descending
+        if (taskCountDiff === 0) {
+          return a.distance - b.distance; // ascending
+        }
+        return taskCountDiff;
+      });
+
+      let routeWasUpdated = false;
+      for (const permutation of orderedFillerTaskCombinationsSorted) {
+        const permutationRoutes = await TaskService.getOptimalRouteForPaths(startTask.location!, [
+          ...permutation.tasks,
+          endTask,
+        ]);
+
+        // Units: seconds
+        const permutationRouteDuration = permutationRoutes.reduce(
+          (acc, route) => acc + route.duration,
+          0
+        );
+
+        // Units: Minutes
+        const permutationTasksDuration = permutation.tasks.reduce(
+          (acc, task) => acc + (task.duration ?? 0),
+          0
+        );
+        const totalDurationSeconds = permutationRouteDuration + permutationTasksDuration * 60;
+        if (totalDurationSeconds * 1000 <= maximumFreeTime && permutationRoutes.length > 0) {
+          // Add permutationRoutes to the updated partialRoutes
+          updatedPartialRoutes.push(...permutationRoutes);
+          updatedCoreTasks.push(...permutation.tasks, endTask);
+          routeWasUpdated = true;
+
+          // Mark the tasks as used
+          for (const task of permutation.tasks) {
+            usedFillerTaskIds.add(task.id);
+          }
+
+          break;
+        }
+      }
+      // Don't squeeze in any filler tasks:
+      if (!routeWasUpdated) {
+        updatedPartialRoutes.push(route);
+        updatedCoreTasks.push(endTask);
+      }
+    }
+    partialRoutes = updatedPartialRoutes;
+    coreTasks = updatedCoreTasks;
+    // END FILLING IN FILLER TASKS
+
+    // START USE REST OF FILLER TASKS
+    const remainingFillerTasks = fillerTasks.filter((task) => !usedFillerTaskIds.has(task.id));
+    if (remainingFillerTasks.length > 0) {
+      const remainingRoute = await TaskService.getOptimalRouteForPaths(
+        coreTasks[coreTasks.length - 1].location!,
+        remainingFillerTasks
+      );
+      partialRoutes.push(...remainingRoute);
+      coreTasks.push(...remainingFillerTasks);
+    }
+    // END USE REST OF FILLER TASKS
+
+    const totalDuration = partialRoutes.reduce((acc, route) => acc + route.duration, 0);
+    const totalDistance = partialRoutes.reduce((acc, route) => acc + route.distance, 0);
+    const segments = partialRoutes.flatMap((route) => route.segments);
+    for (let i = 0; i < segments.length; i++) {
+      segments[i].id = ('segment-' + i).toString();
+    }
+    const taskRoute: Route = {
+      duration: totalDuration,
+      distance: totalDistance,
+      segments: segments,
+    };
+
+    return {
+      route: taskRoute,
+      tasks: coreTasks,
+    };
+  }
+
   static async getOptimalRouteForPaths(startLocation: Location, tasks: Task[]): Promise<Route[]> {
     // Needs to be refactored so that mapbox's api is leveraged to generate the shortest path
     const routePromises: Promise<Route>[] = [];
